refactor(cli): split workspace.rs into a workspace/ module directory
Mechanical split of the 2595-line god-file into focused modules, moved
verbatim with no logic changes:

  mod.rs         module docs, constants, the Store struct, re-exports
  types.rs       domain and view types + status labels
  store.rs       vault/ledger plumbing: open, load, save, check_policy, record
  ops.rs         the business mutations (venture, customers, documents,
                 request/decide/revoke/deliver, drafting assistant)
  kernel_exec.rs the RFC 0003 execution chain + owner secrets + tool manifest
  reporting.rs   command_center, integrity_check, export, guidance derivation
  compose.rs     document rendering and RFC 5322 composition
  verify.rs      offline export verification
  util.rs        input validation, amount parsing, error adapters, clock
  tests.rs       the unchanged test suite

Cross-module internals are pub(super) — nothing new enters the crate's
public surface; ui.rs and main.rs compile unchanged against the same
re-exported API. All 159 tests pass unchanged.
